navyt-api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
yt-dlp==2026.03.17
python-dotenv==1.0.1
httpx==0.27.2
mutagen==1.47.0